Breaks posts page into components
