+++ fact-bounty-client/src/components/PostItem/index.js
@@ -1,0 +1,2 @@
+import PostItem from './PostItem'
+export default PostItem
