+++ fact-bounty-client/src/components/PostsList/index.js
@@ -1,0 +1,2 @@
+import PostsList from './PostsList'
+export default PostsList
