+++ fact-bounty-client/src/components/PostsList/style.sass
@@ -1,0 +1,8 @@
+.post-list-wrapper
+  .postLayout
+    display: grid
+    grid-template-rows: auto
+    min-height: calc(100vh - 64px)
+  .loader
+    text-align: center
+
