+++ fact-bounty-client/src/constants/ApiConstants.js
@@ -1,6 +1,6 @@
 const API_URL_LOCAL = 'http://localhost:5000'
-const API_URL_PROD = 'https://personal-fact-bounty.appspot.com'
+const API_URL_PROD = ''
 
 export const APIConstantsDev = {
-  API_URL: API_URL_PROD
+  API_URL: API_URL_LOCAL
 }
