+++ fact-bounty-client/src/styles/variables.sass
@@ -13,6 +13,9 @@ $text-color: #485460
 $green: #05C46B
 $red: #E01E5A
 $orange: #FFC048
+$mix-color: #0288d1
+$true-color: #009688
+$fake-color: #f4511e
